Haiku: Fix "decimal constant is so large that it is unsigned" warning
Because Haiku error constants are already negative, and ENOMEM is actually
INT32_MIN, the compiler complains about it.

We force the constants to be unsigned before negating them to shut it up.

Tested in both Linux and Haiku, after configuring the resulting values
are the expected ones:

printf("%s: %d\n", #e, OSS_##e)
int main() {
	GEN_ERRNO(E2BIG);
	...
}